feat: 게임시작 입력 유효성 검사 기능 추가
