--- NumberBaseball/NumberBaseball/Extensions/Set+Extensions.swift
@@ -1,14 +1,0 @@
-//
-//  Set+Extensions.swift
-//  NumberBaseball
-//
-//  Created by 허건 on 2022/04/12.
-//
-
-import Foundation
-
-extension Set {
-    func convertArray() -> Array<Int> {
-        return Array(self)
-    }
-}
